I've moved the `saveDescription` action into the base-level `OpportunityController` and removed the `mange.OpportunityController`. "Manage" is a front-end concept, not an api-level one. I've left the route wired to `/manage` for the moment as it will need a coordinated change in the frontend.
