+++ src/main/resources/routes
@@ -8,9 +8,7 @@ GET           /opportunity/:id/application_form                      rifs.busine
 POST          /opportunity/:id/duplicate                             rifs.business.controllers.OpportunityController.duplicate(id: OpportunityId)
 POST          /opportunity/:id/publish                               rifs.business.controllers.OpportunityController.publish(id: OpportunityId)
 
-
-# opportunity management
-POST          /manage/opportunity/:id/description/:sectionNum        rifs.business.controllers.manage.OpportunityController.saveDescription(id: OpportunityId, sectionNum:Int)
+POST          /manage/opportunity/:id/description/:sectionNum        rifs.business.controllers.OpportunityController.saveDescription(id: OpportunityId, sectionNum:Int)
 
 GET           /application_form/:id                                  rifs.business.controllers.ApplicationFormController.byId(id: ApplicationFormId)
